review: fix the Standards-axis findings of the short-export branch
Hard violations:

- Split the Dropbox upload tests. The file went over the 5500-token budget on
  this branch. Its world-building now lives in
  course-publish-dropbox-upload-test-setup.ts, following the shape of
  course-publish-service-test-setup.ts, and the Bundle-reuse tests have their
  own file. Both test files and the setup are inside the budget;
  course-publish-service.ts is the only file the check still names, and it was
  over on main.
- Bring CONTEXT.md up to date with what this branch changed. Exported Video now
  says that a file at the address is not enough — it has to have answered for
  its duration — and Unexported Video now covers the export this machine cannot
  vouch for, not only the one that is missing.
- Say in `cvm course publish --help` (FAILURE HANDLING) that a short encode
  fails the export and therefore the Publish, and that a short file already on
  disk is re-rendered rather than trusted.

Judgement calls:

- Deepen export-sha256-sidecar.ts. `loadExportDigest` both loaded and wrote,
  and `ensureExportDigest` was that same effect with its answer thrown away.
  They are one function under the honest name, `ensureExportDigest`, and
  `computeExportDigest` is now internal.
- `ensureExportDuration` no longer varies over its measure's error type. It
  stays generic in its context alone, which ffprobe genuinely needs.
- Narrow `pauseType` to PauseType once, where the export's Clip durations are
  derived, so no caller casts it back.
- Hand callers of planBundleReuse a ReadonlyMap rather than the Map it built.
- Have scripts/audit-export-durations.mts import the export address and the
  expected duration instead of restating them.
